+++ src/musubi_tuner/minimax_h3_image_train_network.py
@@ -119,7 +119,7 @@ class MiniMaxH3ImageNetworkTrainer(NetworkTrainer):
         encoder = load_minimax_h3_te(
             args.text_encoder,
             device=accelerator.device,
-            compute_dtype=torch.bfloat16,
+            compute_dtype=torch.float32,
             quantize=True,
             tokenizer_dir=args.tokenizer,
             load_mode=args.text_encoder_load_mode,
